LibraryProjectAPI/requirements.txt
asgiref==3.5.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
cryptography==36.0.1
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.4
Django==3.2
django-allauth==0.47.0
django-appconf==1.0.5
django-cors-headers==3.11.0
django-crispy-forms==1.13.0
django-imagekit==4.1.0
django-import-export==2.7.1
django-mathfilters==1.0.0
django-rest-auth==0.9.5
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
et-xmlfile==1.1.0
filelock==3.4.2
idna==3.3
MarkupPy==1.14
mysql-connector-python==8.0.28
mysqlclient==2.1.0
numpy==1.22.3
oauthlib==3.1.1
odfpy==1.4.1
openpyxl==3.0.9
pandas==1.4.2
pilkit==2.0
Pillow==9.0.0
platformdirs==2.4.1
protobuf==3.19.4
pycparser==2.21
PyJWT==2.3.0
PyMySQL==1.0.2
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
PyYAML==6.0
razorpay==1.2.0
requests==2.27.1
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.2
tablib==3.1.0
tzdata==2021.5
urllib3==1.26.8
virtualenv==20.13.0
xlrd==2.0.1
xlwt==1.3.0
